backend/requirements.txt

fastapi
uvicorn
groq
pinecone
supabase
python-dotenv
requests
python-telegram-bot
flake8
black
